Add criteria that error message should be mapped
This is kind of tested through required attribute, but there is a
difference between required state and invalid state. It is also one of
the main reasons to use a specialised input (except switching keyboards
on touch) so it is important these work correctly, and don’t leave
users of AT in dark on why the form hasn’t submitted.

I’ve not included the date, range, and color ones for now, as in Edge
you can’t select an invalid value (at the moment at least). Need to
check if it is needed for these.